Refactor frontend classes out of management commands.
Fixes #23.
