+++ README.md
@@ -31,9 +31,11 @@ Set the BACKUP_BUCKET variable in your settings.py file. This must be a globally
     CARETAKER_ADDITIONAL_BACKUP_PATHS = ['/home/user/path1', '/home/user/path2']
     CARETAKER_BACKEND = 'Amazon S3'
     CARETAKER_BACKENDS = ['caretaker.backend.backends.s3']
+    CARETAKER_FRONTEND = 'Django'
+    CARETAKER_FRONTENDS = ['caretaker.frontend.frontends.django']
     MEDIA_ROOT = '/var/www/media'
 
-The CARETAKER_BACKENDS list allows you to specify the available backends. The CARETAKER_BACKEND variable selects the backend to use (there is only S3 at the moment).
+The CARETAKER_BACKENDS list allows you to specify the available backends. The CARETAKER_BACKEND variable selects the backend to use (there is only S3 at the moment). The same is true of CARETAKER_FRONTENDS and CARETAKER_FRONTEND (which only support Django at the moment).
 
 Generate and run Terraform configuration in your home directory:
